proper hotkeys for layer-related actions
